+++ feature-main/workspace-detail/src/main/java/com/seugi/workspacedetail/feature/workspacedetail/WorkspaceDetailViewModel.kt
@@ -5,8 +5,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.seugi.common.model.Result
 import com.seugi.data.workspace.WorkspaceRepository
-import com.seugi.workspacedetail.feature.model.WorkspaceDetailSideEffect
-import com.seugi.workspacedetail.feature.model.WorkspaceDetailUiState
+import com.seugi.workspacedetail.feature.workspacedetail.model.WorkspaceDetailSideEffect
+import com.seugi.workspacedetail.feature.workspacedetail.model.WorkspaceDetailUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.channels.Channel
